requirements.txt
streamlit
pandas
numpy
seaborn
langchain
openai
google-generativeai
langchain-google-genai
requests
python-dotenv
